# docs/swagger.yaml
basePath: /api
definitions:
...
  serializer.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  serializer.SignUpRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  serializer.SignUpResponse:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: API для управления рецептами, препаратами и расчетом доз.
  license:
    name: MIT
  title: Prescription API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Выполняет аутентификацию и возвращает JWT токен.
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/serializer.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токен доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный логин или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход пользователя
      tags:
      - users
  /users/logout:
    post:
      description: Добавляет текущий JWT токен в blacklist Redis.
      produces:
      - application/json
      responses:
        "204":
          description: Токен добавлен в blacklist
        "401":
          description: Токен не предоставлен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход пользователя
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя системы рецептов.
      parameters:
      - description: Логин и пароль
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/serializer.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/serializer.SignUpResponse'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"